requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
python-multipart
